requirement.txt
pandas
streamlit
plotly-express
nbformat
scipy
